requirements.txt
Django==1.9.7
django-webtest==1.8.0
Pillow==4.0.0
WebTest==2.0.24
gunicorn==19.6.0
dj-database-url==0.4.2
whitenoise==3.2.3
psycopg2==2.6.2
jdatetime==1.8.1
djangorestframework==3.5.4